api_gateway/app/requirements.txt
fastapi
grpcio
grpcio-tools
